Split frame parsing up
This is some precursory work to support multiple PDUs per Ethernet frame.

The EtherCAT header is now split from the PDU frame header, more easily allowing for multiple PDU frames in the future.
