+++ src/pdu_loop/pdu_rx.rs
@@ -1,9 +1,8 @@
 use super::storage::PduStorageRef;
 use crate::{
-    command::Command,
     error::{Error, PduError, PduValidationError},
     fmt,
-    pdu_loop::{frame_header::FrameHeader, pdu_flags::PduFlags},
+    pdu_loop::{frame_header::FrameHeader, pdu_header::PduHeader},
     ETHERCAT_ETHERTYPE, MASTER_ADDR,
 };
 use ethercrab_wire::{EtherCrabWireRead, EtherCrabWireSized};
@@ -52,23 +51,39 @@ impl<'sto> PduRx<'sto> {
 
         let i = raw_packet.payload();
 
-        let header = FramePreamble::unpack_from_slice(i).map_err(|e| {
+        let frame_header = FrameHeader::unpack_from_slice(i).map_err(|e| {
             fmt::error!("Failed to parse frame header: {}", e);
 
             e
         })?;
 
-        let (data, working_counter) = header.data_wkc(i).map_err(|e| {
+        let i = i
+            // Strip EtherCAT frame header...
+            .get(FrameHeader::PACKED_LEN..)
+            // ...then take as much as we're supposed to
+            .and_then(|i| i.get(0..usize::from(frame_header.payload_len)))
+            .ok_or_else(|| {
+                fmt::error!("Received frame is too short");
+
+                Error::ReceiveFrame
+            })?;
+
+        // `i` now contains a complete PDU including header
+
+        // Only support a single PDU per frame for now
+        let pdu_header = PduHeader::unpack_from_slice(i)?;
+
+        let (data, working_counter) = pdu_header.data_wkc(i).map_err(|e| {
             fmt::error!("Could not get frame data/wkc: {}", e);
 
             e
         })?;
 
-        let command = header.command()?;
+        let command = pdu_header.command()?;
 
-        let FramePreamble {
+        let PduHeader {
             index, flags, irq, ..
-        } = header;
+        } = pdu_header;
 
         fmt::trace!(
             "Received frame with index {} ({:#04x}), WKC {}",
@@ -109,295 +124,5 @@ impl<'sto> PduRx<'sto> {
         frame.mark_received(flags, irq, working_counter)?;
 
         Ok(())
-    }
-}
-
-/// PDU frame header, command, index, flags and IRQ.
-#[derive(Debug, Copy, Clone, ethercrab_wire::EtherCrabWireRead)]
-#[wire(bytes = 12)]
-pub struct FramePreamble {
-    #[wire(bytes = 2)]
-    header: FrameHeader,
-
-    // NOTE: The following fields are included in the header length field value.
-    #[wire(bytes = 1)]
-    command_code: u8,
-    #[wire(bytes = 1)]
-    index: u8,
-    #[wire(bytes = 4)]
-    command_raw: [u8; 4],
-    #[wire(bytes = 2)]
-    flags: PduFlags,
-    #[wire(bytes = 2)]
-    irq: u16,
-}
-
-impl FramePreamble {
-    fn data_wkc<'buf>(&self, buf: &'buf [u8]) -> Result<(&'buf [u8], u16), Error> {
-        // Jump past header in the buffer
-        let header_offset = FramePreamble::PACKED_LEN;
-
-        // The length of the PDU data body. There are two bytes after this that hold the working
-        // counter, but are not counted as part of the PDU length from the header.
-        let data_end = usize::from(self.header.payload_len);
-
-        let data = buf.get(header_offset..data_end).ok_or(PduError::Decode)?;
-        let wkc = buf
-            .get(data_end..)
-            .ok_or(Error::Pdu(PduError::Decode))
-            .and_then(|raw| Ok(u16::unpack_from_slice(raw)?))?;
-
-        Ok((data, wkc))
-    }
-
-    fn command(&self) -> Result<Command, Error> {
-        Command::parse_code_data(self.command_code, self.command_raw)
-    }
-
-    /// A hacked equality check used for replay tests only.
-    ///
-    /// It treats `command_raw` specially as this can change in responses.
-    ///
-    /// Please do not use outside the replay tests.
-    #[doc(hidden)]
-    #[allow(unused)]
-    pub fn test_only_hacked_equal(&self, other: &Self) -> bool {
-        self.command_code == other.command_code
-            && self.index == other.index
-            && if matches!(self.command_code, 4 | 5) {
-                self.command_raw == other.command_raw &&
-                self.header == other.header
-            } else {
-                true
-            }
-            // && self.flags == other.flags
-            && self.irq == other.irq
-    }
-
-    /// Similar to [`test_only_hacked_equal`].
-    ///
-    /// Please do not use outside replay tests.
-    #[doc(hidden)]
-    #[allow(unused)]
-    pub fn test_only_hacked_hash(&self, state: &mut impl core::hash::Hasher) {
-        use core::hash::Hash;
-
-        let FramePreamble {
-            header,
-            command_code,
-            index,
-            command_raw,
-            flags: _,
-            irq,
-        } = *self;
-
-        // header.hash(state);
-        command_code.hash(state);
-        index.hash(state);
-
-        if matches!(command_code, 4 | 5) {
-            header.payload_len.hash(state);
-            command_raw.hash(state);
-        }
-
-        // flags.hash(state);
-        irq.hash(state);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::pdu_loop::frame_header::ProtocolType::DlPdu;
-    use core::hash::{Hash, Hasher};
-    use std::collections::{hash_map::DefaultHasher, HashMap};
-
-    // These shouldn't be derived for general use, just for testing
-    impl Eq for FramePreamble {}
-    impl PartialEq for FramePreamble {
-        fn eq(&self, other: &Self) -> bool {
-            self.test_only_hacked_equal(&other)
-        }
-    }
-    impl Hash for FramePreamble {
-        fn hash<H: Hasher>(&self, state: &mut H) {
-            self.test_only_hacked_hash(state);
-        }
-    }
-
-    // Just a sanity check...
-    #[test]
-    fn preamble_eq() {
-        let a = FramePreamble {
-            header: FrameHeader {
-                payload_len: 13,
-                protocol: DlPdu,
-            },
-            command_code: 2,
-            index: 0,
-            command_raw: [0, 0, 0, 0],
-            flags: PduFlags {
-                length: 1,
-                circulated: false,
-                is_not_last: false,
-            },
-            irq: 0,
-        };
-
-        let b = FramePreamble {
-            header: FrameHeader {
-                payload_len: 13,
-                protocol: DlPdu,
-            },
-            command_code: 2,
-            index: 0,
-            command_raw: [0, 0, 0, 0],
-            flags: PduFlags {
-                length: 1,
-                circulated: false,
-                is_not_last: false,
-            },
-            irq: 0,
-        };
-
-        assert_eq!(a, b);
-
-        let mut state = DefaultHasher::new();
-
-        assert_eq!(a.hash(&mut state), b.hash(&mut state));
-    }
-
-    #[test]
-    fn preamble_brd_eq() {
-        let a = FramePreamble {
-            header: FrameHeader {
-                payload_len: 13,
-                protocol: DlPdu,
-            },
-            command_code: 7,
-            index: 0,
-            command_raw: [0, 0, 0, 0],
-            flags: PduFlags {
-                length: 1,
-                circulated: false,
-                is_not_last: false,
-            },
-            irq: 0,
-        };
-
-        let b = FramePreamble {
-            header: FrameHeader {
-                payload_len: 13,
-                protocol: DlPdu,
-            },
-            command_code: 7,
-            index: 0,
-            command_raw: [1, 0, 0, 0],
-            flags: PduFlags {
-                length: 1,
-                circulated: false,
-                is_not_last: false,
-            },
-            irq: 0,
-        };
-
-        // Different `command_raw` but `command_code` is BRD so the equality should still hold.
-        assert_eq!(a, b);
-
-        let mut state_a = DefaultHasher::new();
-        let mut state_b = DefaultHasher::new();
-
-        a.hash(&mut state_a);
-        b.hash(&mut state_b);
-
-        // Hashes remain equal because we look up by sent preamble, not the potentially modified
-        // receive.
-        assert_eq!(state_a.finish(), state_b.finish());
-    }
-
-    #[test]
-    fn find_brd() {
-        let mut map = HashMap::new();
-
-        map.insert(
-            FramePreamble {
-                header: FrameHeader {
-                    payload_len: 13,
-                    protocol: DlPdu,
-                },
-                command_code: 7,
-                index: 0,
-                command_raw: [3, 0, 0, 0],
-                flags: PduFlags {
-                    length: 1,
-                    circulated: false,
-                    is_not_last: false,
-                },
-                irq: 0,
-            },
-            1234usize,
-        );
-
-        assert_eq!(
-            map.get(&FramePreamble {
-                header: FrameHeader {
-                    payload_len: 13,
-                    protocol: DlPdu,
-                },
-                command_code: 7,
-                index: 0,
-                command_raw: [0, 0, 0, 0],
-                flags: PduFlags {
-                    length: 1,
-                    circulated: false,
-                    is_not_last: false,
-                },
-                irq: 0,
-            }),
-            Some(&1234usize)
-        );
-    }
-
-    #[test]
-    fn find_bwr() {
-        let mut map = HashMap::new();
-
-        map.insert(
-            FramePreamble {
-                header: FrameHeader {
-                    payload_len: 14,
-                    protocol: DlPdu,
-                },
-                command_code: 8,
-                index: 1,
-                command_raw: [3, 0, 32, 1],
-                flags: PduFlags {
-                    length: 2,
-                    circulated: false,
-                    is_not_last: false,
-                },
-                irq: 0,
-            },
-            1234usize,
-        );
-
-        assert_eq!(
-            map.get(&FramePreamble {
-                header: FrameHeader {
-                    payload_len: 14,
-                    protocol: DlPdu,
-                },
-                command_code: 8,
-                index: 1,
-                command_raw: [0, 0, 32, 1],
-                flags: PduFlags {
-                    length: 2,
-                    circulated: false,
-                    is_not_last: false,
-                },
-                irq: 0,
-            }),
-            Some(&1234usize)
-        );
     }
 }
